% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{code_message}
\alias{code_message}
\title{API messages}
\format{Codes and corresponding messages from the API.}
\description{
API messages
}
\examples{
data("code_message")
}
\keyword{data}

