% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{countrycodes}
\alias{countrycodes}
\title{Country codes}
\format{
All possible ISO 3166-1 Alpha 2 standard country codes.
}
\description{
Country codes
}
\examples{
data("countrycodes")
}
\keyword{data}
