% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{languagecodes}
\alias{languagecodes}
\title{Language codes}
\format{All possible ISO 639-2 language codes.}
\description{
Language codes
}
\examples{
data("languagecodes")
}
\keyword{data}

