% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_api_ok.R
\name{oc_api_ok}
\alias{oc_api_ok}
\title{Is the OpenCage API available?}
\usage{
oc_api_ok(url = "https://api.opencagedata.com")
}
\arguments{
\item{url}{The URL of the OpenCage API, \url{https://api.opencagedata.com}.}
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}.
}
\description{
Checks whether the OpenCage API can be reached.
}
\examples{
oc_api_ok()
}
