% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_check_key.R
\name{oc_check_key}
\alias{oc_check_key}
\alias{has_oc_key}
\alias{oc_scrub_key}
\alias{oc_show_key}
\title{Check OpenCage API key}
\usage{
oc_check_key(key)

has_oc_key()

oc_scrub_key(string, with = "xxx")

oc_show_key()
}
\arguments{
\item{key}{OpenCage API key}

\item{string}{a url string to be scrubbed, the following values will be scrubbed from the url and replace with the \code{with} argument; key, client, signature}

\item{with}{a string to replace}
}
\description{
This page documents tools for working with your OpenCage API keys. Use \code{oc_config()} to register your OpenCage API key.
}
\details{
Use \code{oc_config()} to register your OpenCage API key. Your key will always be scrubbed in output, the only way to see your actual key is to run \code{oc_show_key()}
}
