% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_config.R
\name{oc_config}
\alias{oc_config}
\title{Set default options}
\usage{
oc_config(max_rate_per_sec = getOption("oc_max_rate_per_sec", default =
  1L), ...)
}
\arguments{
\item{max_rate_per_sec}{maximum number of requests per second}

\item{...}{not used}
}
\description{
Set default options
}
