% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_configure.R
\name{oc_configure}
\alias{oc_configure}
\title{Set default options}
\usage{
oc_configure(max_rate_per_sec = getOption("oc_max_rate_per_sec", default
  = 1L), ...)
}
\arguments{
\item{max_rate_per_sec}{maximum number of requests per second}

\item{...}{not used}
}
\description{
Set default options
}
