% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oc_key}
\alias{oc_key}
\title{Retrieve OpenCage API key}
\usage{
oc_key(quiet = TRUE)
}
\arguments{
\item{quiet}{Logical vector of length one indicating whether the key is
returned quietly or whether a message is printed.}
}
\description{
Retrieves the OpenCage API Key from the environment variable
\code{OPENCAGE_KEY}.
}
