% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_key.R
\name{oc_key_present}
\alias{oc_key_present}
\title{Is an OpenCage API key present?}
\usage{
oc_key_present()
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}.
}
\description{
Checks whether a potential OpenCage API key, i.e. a 32 character long,
alphanumeric string, is stored in the environment variable \code{OPENCAGE_KEY}.
}
\examples{
oc_key_present()
}
