% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oc_query_check}
\alias{oc_query_check}
\title{Query check}
\usage{
oc_query_check(latitude = NULL, longitude = NULL, placename = NULL, key,
  abbrv, bounds, countrycode, language, limit, min_confidence, no_annotations,
  no_dedupe, no_record, add_request)
}
\arguments{
\item{abbrv}{Logical (default FALSE), when TRUE addresses are abbreviated
(e.g. C. instead of Calle)}

\item{bounds}{Provides the geocoder with a hint to the region that
the query resides in. This value will restrict the possible results
to the supplied region. The bounds parameter should be specified as
4 coordinate points forming the south-west and north-east corners of
 a bounding box.
 For example, \code{bounds = c(-0.563160, 51.280430, 0.278970, 51.683979)}
  (min long, min lat, max long, max lat).}

\item{countrycode}{Restricts the results to the given country.
The country code is a two letter code as defined by the ISO 3166-1 Alpha 2
standard. E.g. "GB" for the United Kingdom, "FR" for France,
"US" for United States.}

\item{language}{An IETF format language code
(such as "es" for Spanish or "pt-BR" for Brazilian Portuguese).
 If no language is explicitly specified, we will look for an HTTP
 Accept-Language header like those sent by a browser and use the
 first language specified and if none are specified "en" (English)
  will be assumed.}

\item{limit}{How many results should be returned (1-100). Default is 10.}

\item{min_confidence}{An integer from 1-10. Only results with at least
this confidence will be returned.}

\item{no_annotations}{Logical (default FALSE), when TRUE the output will
not contain annotations.}

\item{no_dedupe}{Logical (default FALSE), when TRUE the output will not be
deduplicated.}

\item{no_record}{Logical (default FALSE), when TRUE no log entry of
the query is created at OpenCage.}

\item{add_request}{Logical (default TRUE), when FALSE the query text
is removed from the results data frame.}
}
\description{
Query check
}
