% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{opencage-deprecated}
\alias{opencage-deprecated}
\title{Deprecated functions in opencage}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{\link[=opencage_forward]{opencage_forward()}}
\item \code{\link[=opencage_reverse]{opencage_reverse()}}
\item \code{\link[=opencage_key]{opencage_key()}}
}
}
