% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_check_query.R
\name{oc_check_query}
\alias{oc_check_query}
\title{Check OpenCage query arguments}
\usage{
oc_check_query(placename = NULL, latitude = NULL, longitude = NULL,
  key = NULL, bounds = NULL, countrycode = NULL, language = NULL,
  limit = NULL, min_confidence = NULL, no_annotations = NULL,
  no_dedupe = NULL, no_record = NULL, abbrv = NULL,
  add_request = NULL)
}
\arguments{
\item{placename}{A character vector with the placename(s) to be geocoded.
Required. See
\href{https://github.com/OpenCageData/opencagedata-misc-docs/blob/master/query-formatting.md}{OpenCage's
 instructions} on how to format addresses for forward geocoding best.}

\item{key}{Your OpenCage API key as a character vector of length one.
Required. By default, \code{\link{oc_key}} will attempt to retrieve the key
from the environment variable \code{OPENCAGE_KEY}.}

\item{bounds}{A list of bounding boxes, i.e. numeric vectors, each with 4
coordinates forming the south-west and north-east corners of a bounding box
\code{list(c(xmin, ymin, xmax, ymax))}. The \code{bounds} parameter will
restrict the possible results to the supplied region. It can easily be
specified with the \code{\link{oc_bbox}} helper, for example like
\code{bounds = oc_bbox(-0.563160, 51.280430, 0.278970, 51.683979)}.}

\item{countrycode}{A two letter code as defined by the
\href{https://www.iso.org/obp/ui/#search/code}{ISO 3166-1 Alpha 2} standard
that restricts the results to the given country or countries. E.g. "AR" for
Argentina, "FR" for France, "NZ" for the New Zealand. Multiple countrycodes
per \code{placename} must be wrapped in a list.}

\item{language}{An
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF language tag}
(such as "es" for Spanish or "pt-BR" for Brazilian Portuguese). OpenCage
will attempt to return results in that language. If it is not specified,
"en" (English) will be assumed by the API.}

\item{limit}{The maximum number of results that should be returned. Integer
values between 1 and 100 are allowed.}

\item{min_confidence}{An integer value between 0 and 10 indicating the
precision of the returned result as defined by it's geographical extent,
(i.e. by the extent of the result's bounding box). See the \href{API
documentation}{https://opencagedata.com/api#confidence} for details. Only
results with at least the requested confidence will be returned.}

\item{no_annotations}{A logical vector indicating whether additional
information about the result location should be returned. \code{TRUE} by
default, which means that the output will not contain annotations.}

\item{no_dedupe}{A logical vector (default \code{FALSE}), when \code{TRUE}
the output will not be deduplicated.}

\item{no_record}{A logical vector (default \code{FALSE}), when \code{TRUE} no
log entry of the query is created and the geocoding request is not cached
by OpenCage.}

\item{abbrv}{A logical vector (default \code{FALSE}), when \code{TRUE}
addresses in the \code{formatted} field of the results are abbreviated
(e.g. "Main St." instead of "Main Street").}

\item{add_request}{A logical vector (default \code{FALSE}) indicating whether
the request is returned again with the results. If the \code{return} value
is a \code{df_list}, the query text is added as a column to the results.
\code{json_list} results will contain all request parameters, including the
API key used! For \code{geojson_list} this is currently ignored by
OpenCage.}
}
\description{
Function that checks the query arguments passed to OpenCage
}
\keyword{internal}
