% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{oc_clear_cache}
\alias{oc_clear_cache}
\title{Clear the opencage cache}
\usage{
oc_clear_cache()
}
\description{
Forget past results and reset the \pkg{opencage} cache.
}
\examples{
\dontshow{if (oc_key_present() && oc_api_ok()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

system.time(oc_reverse(latitude = 10, longitude = 10))
system.time(oc_reverse(latitude = 10, longitude = 10))
oc_clear_cache()
system.time(oc_reverse(latitude = 10, longitude = 10))
\dontshow{\}) # examplesIf}
}
