% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_config.R
\name{oc_config}
\alias{oc_config}
\title{Configure settings}
\usage{
oc_config(key = Sys.getenv("OPENCAGE_KEY"),
  rate_sec = getOption("oc_rate_sec", default = 1L), ...)
}
\arguments{
\item{key}{Your OpenCage API key as a character vector of length one. Do not
pass the key directly as a parameter, though. See details.}

\item{rate_sec}{Numeric vector of length one. Sets the maximum number of
requests sent to the OpenCage API per second. Defaults to the value set in
the \code{oc_rate_sec} option, or, in case that does not exist, to 1L.}

\item{...}{Ignored.}
}
\description{
Configure session settings for \pkg{opencage}.
}
\section{Set your OpenCage API key}{


\pkg{opencage} uses the environment variable \code{OPENCAGE_KEY} internally
to retrieve your OpenCage key. \code{\link{oc_config}} will help to set that
environment variable. Do not pass the key directly as a parameter to the
function, though, as you risk exposing it via your script or your history.
There are three safer ways to set your API key instead:

1. Save your API key as an environment variable in
\code{\link[base:Startup]{.Renviron}} as described in
\href{https://rstats.wtf/r-startup.html#renviron}{What They Forgot to Teach
You About R} or
\href{https://csgillespie.github.io/efficientR/set-up.html#renviron}{Efficient
R Programming}. From there it will be fetched by all functions that call the
OpenCage API, so you do not even have to call \code{oc_config} to set your
key, but can start geocoding right away. If you have the \pkg{usethis}
package installed, you can edit your \code{\link[base:Startup]{.Renviron}}
most easily with \code{usethis::edit_r_environ()}.

2. If you use a package like \pkg{keyring} to store your credentials, you can
safely pass your key in a script with a function call like this
\code{oc_config(key = keyring::key_get("opencage"))}.

3. If you call \code{oc_config} in an \code{\link[base]{interactive}} session and the
\code{OPENCAGE_KEY} environment variable is not set, it will prompt you to
enter the key in the console.
}

\section{Set your OpenCage API rate limit}{


The rate limit allowed by the API depends on the OpenCage plan you
purchased and ranges from 1 request/sec for the "Free Trial" plan to 15
requests/sec for the "Medium" or "Large" plans, see
\url{https://opencagedata.com/pricing} for details and up-to-date
information. You can set the rate limit persistently across sessions by
setting the \code{oc_rate_sec} \link[base:options]{option} in your
\code{\link[base:Startup]{.Rprofile}}. If you have the \code{usethis}
package installed, you can edit your \code{\link[base:Startup]{.Rprofile}}
most easily with \code{usethis::edit_r_profile()}.
}

