% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_forward.R
\name{oc_forward_df}
\alias{oc_forward_df}
\title{Forward geocoding with data frames}
\usage{
oc_forward_df(data, placename, bind_cols = TRUE, output = c("short",
  "all"), key = oc_key(), bounds = NULL, countrycode = NULL,
  language = NULL, limit = 1L, min_confidence = NULL,
  no_annotations = TRUE, no_dedupe = FALSE, no_record = FALSE,
  abbrv = FALSE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{placename}{An unquoted variable name of a character vector with the
  placenames to be geocoded.

  If the placenames are addresses, see
  \href{https://github.com/OpenCageData/opencagedata-misc-docs/blob/master/query-formatting.md}{OpenCage's
   instructions} on how to format addresses for best forward geocoding
   results.}

\item{bind_cols}{When \code{bind_col = TRUE}, the default, the results are
column bound to \code{data}. When \code{FALSE}, the results are returned as
a new tibble.}

\item{output}{A character vector of length one indicating whether only
latitude, longitude, and formatted address variables (\code{"short"}, the
default) should be returned or all variables (\code{"all"}) variables
should be returned.}

\item{key}{Your OpenCage API key as a character vector of length one. By
default, \code{\link{oc_key}} will attempt to retrieve the key from the
environment variable \code{OPENCAGE_KEY}.}

\item{bounds}{A list, or an unquoted variable name of a list column of
bounding boxes, i.e. a list of numeric vectors, each with 4 coordinates
forming the south-west and north-east corners of a bounding box:
\code{list(c(xmin, ymin, xmax, ymax))}. \code{bounds} restricts the
possible results to the supplied region. It can be specified with the
\code{\link{oc_bbox}} helper. For example: \code{bounds =
oc_bbox(-0.563160, 51.280430, 0.278970, 51.683979)}. Default is
\code{NULL}.}

\item{countrycode}{Character vector, or an unquoted variable name of such a
vector, of two-letter codes as defined by the
\href{https://www.iso.org/obp/ui/#search/code}{ISO 3166-1 Alpha 2} standard
that restricts the results to the given country or countries. E.g. "AR" for
Argentina, "FR" for France, "NZ" for the New Zealand. Multiple countrycodes
per \code{placename} must be wrapped in a list. Default is \code{NULL}.}

\item{language}{Character vector, or an unquoted varible name of such a
vector, of
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF language tags}
(such as "es" for Spanish or "pt-BR" for Brazilian Portuguese). OpenCage
will attempt to return results in that language. If it is not specified,
"en" (English) will be assumed by the API.}

\item{limit}{Numeric vector of integer values, or an unquoted variable name
of such a vector, to determine the maximum number of results returned for
each \code{placename}. Integer values between 1 and 100 are allowed.
Default is 1.}

\item{min_confidence}{Numeric vector of integer values, or an unquoted
variable name of such a vector, between 0 and 10 indicating the precision
of the returned result as defined by its geographical extent, (i.e. by the
extent of the result's bounding box). See the
\href{https://opencagedata.com/api#confidence}{API documentation} for
details. Only results with at least the requested confidence will be
returned. Default is \code{NULL}).}

\item{no_annotations}{Logical vector, or an unquoted variable name of such a
vector, indicating whether additional information about the result location
should be returned. Default is \code{TRUE}, which means that the output
will not contain annotations.}

\item{no_dedupe}{Logical vector, or an unquoted variable name of such a
vector. Default is \code{FALSE}. When \code{TRUE} the output will not be
deduplicated.}

\item{no_record}{Logical vector (default \code{FALSE}), when \code{TRUE} no
log entry of the query is created, and the geocoding request is not cached
by OpenCage.}

\item{abbrv}{Logical vector, or an unquoted variable name of such a
vector. Default is \code{FALSE}. When \code{TRUE} addresses in the
\code{formatted} variable of the results are abbreviated (e.g. "Main St."
instead of "Main Street").}

\item{...}{Ignored.}
}
\value{
A tibble.
}
\description{
Forward geocoding from a placename variable to latitude and longitude
  tuple(s).
}
\examples{

\dontrun{
library(tibble)
df <- tibble(id = 1:3,
             locations = c("Nantes", "Hamburg", "Los Angeles"))

# Return lat, lng, and formatted address
oc_forward_df(df, placename = locations)

# Return more detailed information about the locations
oc_forward_df(df, placename = locations, output = "all")

# Do not column bind results to input data frame
oc_forward_df(df, placename = locations, bind_cols = FALSE)

# Add more results by changing the limit from the default of 1.
oc_forward_df(df, placename = locations, limit = 5)

# Restrict results to a given bounding box
oc_forward_df(df, placename = locations,
              bounds = oc_bbox(-5, 45, 15, 55))

# oc_forward_df accepts unquoted column names for all
# arguments except bind_cols, output, key, and no_record.
# This makes it possible to build up more detailed queries
# through the data frame passed to the data argument.

df2 <- add_column(df,
  bounds = oc_bbox(xmin = c(-2, 9, -119),
                   ymin = c(47, 53, 34),
                   xmax = c(0, 10, -117),
                   ymax = c(48, 54, 35)),
  limit = 1:3,
  countrycode = c("ca", "us", "co"),
  language = c("fr", "de", "en"))

# Use the bounds column to help return accurate results and
# language column to specify preferred language of results
oc_forward_df(df2, placename = locations,
              bounds = bounds,
              language = language)

# Different limit of results for each placename
oc_forward_df(df2, placename = locations,
              limit = limit)

# Specify the desired results by the countrycode column
oc_forward_df(df2, placename = locations,
              countrycode = countrycode)
}
}
\seealso{
\code{\link{oc_forward}} for inputs as vectors, or
  \code{\link{oc_reverse}} and \code{\link{oc_reverse_df}} for reverse
  geocoding. For more information about the API and the various parameters,
  see the \href{https://opencagedata.com/api}{OpenCage API documentation}.
}
