% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_reverse.R
\name{oc_reverse}
\alias{oc_reverse}
\alias{oc_reverse_df}
\title{Reverse geocoding}
\usage{
oc_reverse(latitude, longitude, output = c("df_list", "json_list",
  "geojson_list", "url_only"), key = oc_key(), language = NULL,
  limit = 10L, min_confidence = NULL, no_annotations = TRUE,
  no_dedupe = FALSE, no_record = FALSE, abbrv = FALSE,
  add_request = FALSE, ...)

oc_reverse_df(data, latitude, longitude, bind_cols = TRUE,
  output = c("short", "all"), key = oc_key(), language = NULL,
  limit = 1L, min_confidence = NULL, no_annotations = TRUE,
  no_dedupe = FALSE, no_record = FALSE, abbrv = FALSE, ...)
}
\arguments{
\item{latitude}{Latitude.}

\item{longitude}{Longitude.}

\item{output}{The return value of the function, either a \code{df_list}, a \code{json_list}, a \code{geojson_list}, or \code{url_only.}}

\item{key}{Your OpenCage key.}

\item{language}{An IETF format language code (such as "es" for Spanish or
"pt-BR" for Brazilian Portuguese). If no language is explicitly specified,
we will look for an HTTP Accept-Language header like those sent by a
browser and use the first language specified and if none are specified "en"
(English) will be assumed.}

\item{limit}{How many results should be returned (1-100)? Default is 10.}

\item{min_confidence}{An integer from 1-10. Only results with at least this
confidence will be returned.}

\item{no_annotations}{Logical (default FALSE), when TRUE the output will not
contain annotations.}

\item{no_dedupe}{Logical (default FALSE), when TRUE the output will not be
deduplicated.}

\item{no_record}{Logical (default FALSE), when TRUE no log entry of the query
is created at OpenCage.}

\item{abbrv}{Logical (default FALSE), when TRUE addresses are abbreviated
(e.g. C. instead of Calle)}

\item{add_request}{Logical (default TRUE), when FALSE the query text is
removed from the results data frame.}

\item{...}{Currently not used.}

\item{data}{A data frame}

\item{bind_cols}{logical Bind source and results data frame?}
}
\value{
Depending on the \code{output} argument list of data.frame(s) (\code{df_list})
}
\description{
Reverse geocoding, from latitude and longitude to placename(s).
}
\details{
\strong{API key}
To get an API key to access OpenCage geocoding,
register at \url{https://geocoder.opencagedata.com/pricing}.
The free API key provides up to 2,500 calls a day. For ease of use,
 save your API key as an environment variable as described at
  \url{https://stat545-ubc.github.io/bit003_api-key-env-var.html}.
Both functions of the package will conveniently look for your API key
using \code{Sys.getenv("OPENCAGE_KEY")} so if your API key is an environment
 variable called "OPENCAGE_KEY" you don't need to input it manually.

\strong{memoise}
The underlying data at OpenCage is updated about once a day.
Note that the package uses `memoise` with no timeout argument so that results
 are cached inside an active R session.

This function typically returns multiple results due to placename ambiguity;
 consider using the \code{bounds} parameter to limit the area searched.
}
\examples{
\dontrun{
oc_reverse(latitude = 0, longitude = 0, limit = 2)
}
}
