% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencage-package.R
\docType{package}
\name{opencage-package}
\alias{opencage}
\alias{opencage-package}
\title{opencage: Geocode with the OpenCage API}
\description{
Geocode with the OpenCage API, either from place name to longitude and latitude (forward geocoding) or from longitude and latitude to the name and address of a location (reverse geocoding), see \url{https://opencagedata.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/opencage/}
  \item \url{https://github.com/ropensci/opencage}
  \item Report bugs at \url{https://github.com/ropensci/opencage/issues}
}

}
\author{
\strong{Maintainer}: Daniel Possenriede \email{possenriede+r@gmail.com} (\href{https://orcid.org/0000-0002-6738-9845}{ORCID})

Authors:
\itemize{
  \item Jesse Sadler (\href{https://orcid.org/0000-0001-6081-9681}{ORCID})
  \item Maëlle Salmon (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

Other contributors:
\itemize{
  \item Noam Ross [contributor]
  \item Jake Russ [contributor]
  \item Julia Silge (Julia Silge reviewed \{opencage\} (v. 0.1.0) for rOpenSci, see <https://github.com/ropensci/onboarding/issues/36>.) [reviewer]
}

}
\keyword{internal}
