% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{opencage_key}
\alias{opencage_key}
\title{Retrieve Opencage API key}
\usage{
opencage_key(quiet = TRUE)
}
\arguments{
\item{quiet}{Logical vector of length one indicating whether the key is
returned quietly or whether a message is printed.}
}
\description{
\lifecycle{soft-deprecated}

Soft-deprecated and will be removed from the package together with
\code{opencage_forward()} and \code{opencage_reverse()}.

Retrieves the OpenCage API Key from the environment variable \code{OPENCAGE_KEY}.
}
\keyword{internal}
